util: Re-use `ShellKind` in environment sourcing
